Move decorator actions to JTree class
And clean up several parts of the code because the more I write the clearer it becomes to me how this should be structured in a clean way.
