+++ src/main/kotlin/com/fwdekker/randomness/template/TemplateListEditor.kt
@@ -1,6 +1,5 @@
 package com.fwdekker.randomness.template
 
-import com.fwdekker.randomness.ActuallyPrivate
 import com.fwdekker.randomness.Scheme
 import com.fwdekker.randomness.SettingsState
 import com.fwdekker.randomness.StateEditor
@@ -18,22 +17,8 @@ import com.fwdekker.randomness.uuid.UuidScheme
 import com.fwdekker.randomness.uuid.UuidSchemeEditor
 import com.fwdekker.randomness.word.WordScheme
 import com.fwdekker.randomness.word.WordSchemeEditor
-import com.intellij.icons.AllIcons
-import com.intellij.openapi.actionSystem.ActionToolbarPosition
-import com.intellij.openapi.actionSystem.AnAction
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.ui.popup.JBPopupFactory
-import com.intellij.openapi.ui.popup.ListSeparator
-import com.intellij.openapi.ui.popup.PopupStep
-import com.intellij.openapi.ui.popup.util.BaseListPopupStep
-import com.intellij.ui.AnActionButton
-import com.intellij.ui.CommonActionsPanel
 import com.intellij.ui.JBSplitter
-import com.intellij.ui.LayeredIcon
-import com.intellij.ui.ToolbarDecorator
 import com.intellij.ui.components.JBScrollPane
-import com.intellij.ui.treeStructure.Tree
-import com.intellij.util.ui.JBUI
 import java.awt.BorderLayout
 import javax.swing.JPanel
 import javax.swing.SwingUtilities
@@ -55,7 +40,7 @@ import javax.swing.SwingUtilities
 class TemplateListEditor(settings: SettingsState = SettingsState.default) : StateEditor<SettingsState>(settings) {
     override val rootComponent = JPanel(BorderLayout())
     private var currentState: SettingsState = SettingsState()
-    private val templateTree = TemplateJTree { isModified(it) }
+    private val templateTree = TemplateJTree(originalState, currentState)
     private var schemeEditorPanel = JPanel(BorderLayout())
     private var schemeEditor: StateEditor<*>? = null
 
@@ -74,7 +59,7 @@ class TemplateListEditor(settings: SettingsState = SettingsState.default) : Stat
 
         // Left half
         templateTree.addTreeSelectionListener { onTreeSelection() }
-        splitter.firstComponent = JBScrollPane(decorateTemplateList(templateTree))
+        splitter.firstComponent = JBScrollPane(templateTree.asDecoratedPanel())
 
         // Right half
         val previewPanel = PreviewPanel {
@@ -91,38 +76,6 @@ class TemplateListEditor(settings: SettingsState = SettingsState.default) : Stat
 
         loadState()
     }
-
-
-    /**
-     * Decorates the given list with buttons for adding, removing, copying, etc.
-     *
-     * @param templateList the list to decorate
-     * @return a panel containing both the decorator and the given list
-     */
-    private fun decorateTemplateList(templateList: Tree) =
-        // TODO: Move this method and button definitions to [TemplateJTree]
-        ToolbarDecorator.createDecorator(templateList)
-            .setToolbarPosition(ActionToolbarPosition.TOP)
-            .setPanelBorder(JBUI.Borders.empty())
-            .setScrollPaneBorder(JBUI.Borders.empty())
-            .disableAddAction()
-            .disableRemoveAction()
-            .addExtraAction(AddSchemeActionButton())
-            .addExtraAction(RemoveActionButton())
-            .addExtraAction(CopyActionButton())
-            .addExtraAction(UpActionButton())
-            .addExtraAction(DownActionButton())
-            .addExtraAction(ResetActionButton())
-            .setButtonComparator("Add", "Edit", "Remove", "Copy", "Up", "Down", "Reset")
-            .createPanel()
-            .also {
-                // Focus on editor when `Enter` is pressed
-                object : AnAction() {
-                    override fun actionPerformed(event: AnActionEvent) {
-                        schemeEditor?.preferredFocusedComponent?.requestFocus()
-                    }
-                }.registerCustomShortcutSet(CommonActionsPanel.getCommonShortcut(CommonActionsPanel.Buttons.EDIT), it)
-            }
 
     /**
      * Invoked when an entry is (de)selected in the tree.
@@ -172,32 +125,11 @@ class TemplateListEditor(settings: SettingsState = SettingsState.default) : Stat
         }
 
 
-    /**
-     * Adds the given scheme to the tree.
-     *
-     * @param newScheme the scheme to add
-     * @see TemplateJTree.addScheme
-     */
-    @ActuallyPrivate("Exposed for testing because popup cannot easily be tested.")
-    internal fun addScheme(newScheme: Scheme) = templateTree.addScheme(newScheme)
-
-    /**
-     * Returns true if and only if the given scheme has been modified with respect to [originalState].
-     *
-     * @param scheme the scheme to check for modification
-     * @return true if and only if the given scheme has been modified with respect to [originalState]
-     */
-    private fun isModified(scheme: Scheme) =
-        originalState.templateList.getSchemeByUuid(scheme.uuid) != scheme ||
-            scheme is StringScheme && originalState.symbolSetSettings != currentState.symbolSetSettings ||
-            scheme is WordScheme && originalState.dictionarySettings != currentState.dictionarySettings
-
-
     override fun loadState(state: SettingsState) {
         super.loadState(state)
 
         currentState.copyFrom(state)
-        templateTree.reload(currentState.templateList)
+        templateTree.reload()
     }
 
     override fun readState() = currentState.deepCopy(retainUuid = true)
@@ -205,8 +137,8 @@ class TemplateListEditor(settings: SettingsState = SettingsState.default) : Stat
     override fun reset() {
         super.reset()
 
-        if (queueSelection != null) {
-            templateTree.selectScheme(queueSelection)
+        queueSelection?.also {
+            templateTree.selectedScheme = currentState.templateList.getSchemeByUuid(it)
             SwingUtilities.invokeLater { schemeEditor?.preferredFocusedComponent?.requestFocus() }
 
             queueSelection = null
@@ -217,129 +149,6 @@ class TemplateListEditor(settings: SettingsState = SettingsState.default) : Stat
     override fun addChangeListener(listener: () -> Unit) {
         templateTree.model.addTreeModelListener(SimpleTreeModelListener(listener))
         templateTree.addTreeSelectionListener { listener() }
-    }
-
-
-    /**
-     * Displays a popup to add a scheme, or immediately adds a template if nothing is currently selected.
-     */
-    private inner class AddSchemeActionButton : AnActionButton("Add", AllIcons.General.Add) {
-        override fun actionPerformed(event: AnActionEvent) {
-            if (templateTree.selectedNodeNotRoot == null) {
-                addScheme(AVAILABLE_ADD_SCHEMES[0])
-                return
-            }
-
-            JBPopupFactory.getInstance()
-                .createListPopup(AddSchemePopupStep())
-                .show(preferredPopupPoint ?: return)
-        }
-
-        override fun getShortcut() = CommonActionsPanel.getCommonShortcut(CommonActionsPanel.Buttons.ADD)
-
-        override fun updateButton(event: AnActionEvent) {
-            super.updateButton(event)
-
-            event.presentation.icon =
-                if (templateTree.selectedNodeNotRoot == null) AllIcons.General.Add
-                else LayeredIcon.ADD_WITH_DROPDOWN
-        }
-
-
-        /**
-         * The [Scheme]-related entries in [AddSchemeActionButton].
-         */
-        private inner class AddSchemePopupStep : BaseListPopupStep<Scheme>(null, AVAILABLE_ADD_SCHEMES) {
-            override fun getIconFor(value: Scheme?) = value?.icon
-
-            override fun getTextFor(value: Scheme?) = value?.name ?: Scheme.DEFAULT_NAME
-
-            override fun onChosen(value: Scheme?, finalChoice: Boolean): PopupStep<*>? {
-                if (value != null)
-                    templateTree.addScheme(value.deepCopy().also { it.setSettingsState(currentState) })
-
-                return null
-            }
-
-            override fun isSpeedSearchEnabled() = true
-
-            override fun getSeparatorAbove(value: Scheme?) =
-                if (value == AVAILABLE_ADD_SCHEMES[1]) ListSeparator()
-                else null
-
-            override fun getDefaultOptionIndex() = 0
-        }
-    }
-
-    /**
-     * Removes the selected scheme.
-     */
-    private inner class RemoveActionButton : AnActionButton("Remove", AllIcons.General.Remove) {
-        override fun actionPerformed(event: AnActionEvent) = templateTree.removeScheme(templateTree.selectedScheme!!)
-
-        override fun isEnabled() = templateTree.selectedNodeNotRoot != null
-
-        override fun getShortcut() = CommonActionsPanel.getCommonShortcut(CommonActionsPanel.Buttons.REMOVE)
-    }
-
-    /**
-     * Duplicates the selected scheme.
-     */
-    private inner class CopyActionButton : AnActionButton("Copy", AllIcons.Actions.Copy) {
-        override fun actionPerformed(event: AnActionEvent) {
-            val copy = templateTree.selectedScheme!!.deepCopy()
-            copy.setSettingsState(currentState)
-
-            templateTree.addScheme(copy)
-        }
-
-        override fun isEnabled() = templateTree.selectedNodeNotRoot != null
-    }
-
-    /**
-     * Moves a scheme up by one position.
-     */
-    private inner class UpActionButton : AnActionButton("Up", AllIcons.Actions.MoveUp) {
-        override fun actionPerformed(event: AnActionEvent) =
-            templateTree.moveSchemeDownBy(templateTree.selectedScheme!!, -1)
-
-        override fun isEnabled() =
-            templateTree.selectedScheme?.let { templateTree.canMoveSchemeDownBy(it, -1) } ?: false
-
-        override fun getShortcut() = CommonActionsPanel.getCommonShortcut(CommonActionsPanel.Buttons.UP)
-    }
-
-    /**
-     * Moves a scheme down by one position.
-     */
-    private inner class DownActionButton : AnActionButton("Down", AllIcons.Actions.MoveDown) {
-        override fun actionPerformed(event: AnActionEvent) =
-            templateTree.moveSchemeDownBy(templateTree.selectedScheme!!, 1)
-
-        override fun isEnabled() =
-            templateTree.selectedScheme?.let { templateTree.canMoveSchemeDownBy(it, 1) } ?: false
-
-        override fun getShortcut() = CommonActionsPanel.getCommonShortcut(CommonActionsPanel.Buttons.DOWN)
-    }
-
-    /**
-     * Resets the selected scheme to its original state, or removes it if it has no original state.
-     */
-    private inner class ResetActionButton : AnActionButton("Reset", AllIcons.General.Reset) {
-        override fun actionPerformed(event: AnActionEvent) {
-            val toReset = templateTree.selectedScheme!!
-            val toResetFrom = originalState.templateList.getSchemeByUuid(toReset.uuid)
-            if (toResetFrom == null) {
-                templateTree.removeScheme(toReset)
-                return
-            }
-
-            toReset.copyFrom(toResetFrom)
-            toReset.setSettingsState(currentState)
-            templateTree.selectScheme(toReset.uuid)
-        }
-
-        override fun isEnabled() = templateTree.selectedScheme?.let { isModified(it) } ?: false
     }
 
 
@@ -356,25 +165,5 @@ class TemplateListEditor(settings: SettingsState = SettingsState.default) : Stat
          * The default proportion of the splitter component.
          */
         const val DEFAULT_SPLITTER_PROPORTION = .25f
-
-        /**
-         * The text that is displayed when the table is empty.
-         */
-        const val EMPTY_TEXT = "No templates configured."
-
-        /**
-         * Returns the list of schemes that the user can add from the add action.
-         */
-        val AVAILABLE_ADD_SCHEMES: List<Scheme>
-            get() = listOf(
-                Template("Template", emptyList()),
-                IntegerScheme(),
-                DecimalScheme(),
-                StringScheme(),
-                WordScheme(),
-                UuidScheme(),
-                LiteralScheme(),
-                TemplateReference()
-            )
     }
 }
