+++ src/test/kotlin/com/fwdekker/randomness/template/TemplateListEditorTest.kt
@@ -5,7 +5,9 @@ import org.spekframework.spek2.Spek
 
 /**
  * GUI tests for [TemplateListEditor].
+ *
+ * TODO: Docs: Unit or integration tests?
  */
 object TemplateListEditorTest : Spek({
-    TODO()
+    //
 })
